+++ stats/mock_collector.go
@@ -76,5 +76,5 @@ type MockMetric struct {
 }
 
 func millis(d time.Duration) float64 {
-	return float64(d / time.Millisecond)
+	return float64(d) / float64(time.Millisecond)
 }
